NewTemplateSignature
